Drop MUT* macros and C support.
The MUT* macros were only used in one place, and the 0 in
MUTCONSTRUCT(0) meant that the mutex was never actually used.
