+++ coreneuron/io/nrn_setup.cpp
@@ -20,7 +20,7 @@
 #include "coreneuron/sim/fast_imem.hpp"
 #include "coreneuron/network/multisend.hpp"
 #include "coreneuron/utils/nrn_assert.h"
-#include "coreneuron/utils/nrnmutdec.h"
+#include "coreneuron/utils/nrnmutdec.hpp"
 #include "coreneuron/utils/memory.h"
 #include "coreneuron/utils/utils.hpp"
 #include "coreneuron/mpi/nrnmpi.h"
